+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 :warning: There could be breaking changes anywhere during major version zero(v0.x.x).
 
+## v0.1.1 (Unreleased)
+
 ## v0.1.0 (Unreleased)
 
 ### Breaking changes